+++ wsgi/settings.py
@@ -1,0 +1,8 @@
+DOCUMENT_ROOT = "./resource/server"
+
+MIME_TYPES = {
+    ".txt": b"text/plain",
+    ".html": b"text/html; charset=UTF-8",
+    ".css": b"text/css",
+    ".js": b"text/javascript",
+}
